+++ insets/src/main/java/dev/chrisbanes/accompanist/insets/Size.kt
@@ -21,9 +21,7 @@
 
 package dev.chrisbanes.accompanist.insets
 
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.height
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.LayoutModifier
 import androidx.compose.ui.Measurable
 import androidx.compose.ui.MeasureScope
@@ -31,12 +29,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.composed
 import androidx.compose.ui.layout.IntrinsicMeasurable
 import androidx.compose.ui.layout.IntrinsicMeasureScope
-import androidx.compose.ui.platform.DensityAmbient
-import androidx.compose.ui.platform.LayoutDirectionAmbient
 import androidx.compose.ui.unit.Constraints
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 
 /**
@@ -147,44 +142,6 @@ fun Modifier.navigationBarsWidth(
         insets = AmbientWindowInsets.current.navigationBars,
         widthSide = side,
         additionalWidth = additional
-    )
-}
-
-/**
- * Returns the current insets converted into a [PaddingValues].
- *
- * @param start Whether to apply the inset on the start dimension.
- * @param top Whether to apply the inset on the top dimension.
- * @param end Whether to apply the inset on the end dimension.
- * @param bottom Whether to apply the inset on the bottom dimension.
- */
-@Composable
-fun Insets.toPaddingValues(
-    start: Boolean = true,
-    top: Boolean = true,
-    end: Boolean = true,
-    bottom: Boolean = true
-): PaddingValues = with(DensityAmbient.current) {
-    val layoutDirection = LayoutDirectionAmbient.current
-    PaddingValues(
-        start = when {
-            start && layoutDirection == LayoutDirection.Ltr -> this@toPaddingValues.left.toDp()
-            start && layoutDirection == LayoutDirection.Rtl -> this@toPaddingValues.right.toDp()
-            else -> 0.dp
-        },
-        top = when {
-            top -> this@toPaddingValues.top.toDp()
-            else -> 0.dp
-        },
-        end = when {
-            end && layoutDirection == LayoutDirection.Ltr -> this@toPaddingValues.right.toDp()
-            end && layoutDirection == LayoutDirection.Rtl -> this@toPaddingValues.left.toDp()
-            else -> 0.dp
-        },
-        bottom = when {
-            bottom -> this@toPaddingValues.bottom.toDp()
-            else -> 0.dp
-        }
     )
 }
